refactor: address PR feedback for comment components
- Move SpeechBubbleIcon to src/components/svg
- Create CenteredWrapper utility component in src/components/core/common
- Move Comment interface to src/types.ts for reusability
- Remove test-comments page as requested
- Integrate CommentsList into ProfilePage with mock data
