+++ src/components/Dashboard/Profile/ProfilePage.tsx
@@ -7,6 +7,8 @@ import { ApiVolunteerGet } from "need4deed-sdk";
 // import { PROFILE_CARD_CONFIG } from "./config/ProfileSectionConfig";
 import { useTranslation } from "react-i18next";
 import { VolunteerHeader } from "./sections/VolunteerHeader";
+import { CommentsList } from "@/components/core/comment";
+import { Comment } from "@/types";
 
 const PageContainer = styled.div`
   max-width: 1200px;
@@ -59,6 +61,44 @@ interface ProfilePageProps {
 
 const ProfilePage = ({ volunteer }: ProfilePageProps) => {
   const { t } = useTranslation();
+
+  // Mock comments data - replace with real API call
+  const mockComments: Comment[] = [
+    {
+      id: "1",
+      userId: "user1",
+      userName: "Kalima",
+      text: "Leni works as a social worker with special-needs children, supporting learning development",
+      timestamp: new Date("2024-08-04T13:14:00").toISOString(),
+    },
+    {
+      id: "2",
+      userId: "user2",
+      userName: "Kalima",
+      text: "Comment example here. An emergency with the appointment.",
+      timestamp: new Date("2024-09-01T12:12:00").toISOString(),
+    },
+  ];
+
+  // Mock functions - replace with real API calls
+  const handlePatch = async (commentId: string, updatedText: string) => {
+    console.log("PATCH:", commentId, updatedText);
+    // TODO: Implement real API call
+    await new Promise((resolve) => setTimeout(resolve, 500));
+  };
+
+  const handleDelete = async (commentId: string) => {
+    console.log("DELETE:", commentId);
+    // TODO: Implement real API call
+    await new Promise((resolve) => setTimeout(resolve, 500));
+  };
+
+  const handleAdd = async (text: string) => {
+    console.log("ADD:", text);
+    // TODO: Implement real API call
+    await new Promise((resolve) => setTimeout(resolve, 500));
+  };
+
   return (
     <PageContainer>
       <BackLink href="/dashboard">
@@ -77,6 +117,15 @@ const ProfilePage = ({ volunteer }: ProfilePageProps) => {
 
       <Card>
         <VolunteerHeader volunteer={volunteer} />
+      </Card>
+
+      <Card>
+        <CommentsList
+          comments={mockComments}
+          onPatch={handlePatch}
+          onDelete={handleDelete}
+          onAdd={handleAdd}
+        />
       </Card>
     </PageContainer>
   );
