+++ src/components/core/common/CenteredWrapper.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const CenteredWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
+export default CenteredWrapper;
+
