+++ src/components/core/common/index.ts
@@ -1,3 +1,4 @@
 export * from "./Search";
 export * from "./Tags";
 export * from "./FormInput";
+export { default as CenteredWrapper } from "./CenteredWrapper";
