+++ src/components/svg/SpeechBubbleIcon.tsx
