+++ src/types.ts
@@ -55,3 +55,11 @@ export type LanguageObject = {
 export type FooterLink = [string, string];
 
 export type HttpMethod = "post" | "patch" | "put";
+
+export interface Comment {
+  id: string;
+  userId: string;
+  userName: string;
+  text: string;
+  timestamp: string;
+}
